wip: prototype a tullia clade
During implementation, I noticed a couple of challenge:

- The data tranformer is a bit wild :-)
- Bins can't be lazily built as part of the RunSpec
- Dependencies within tullia would have to be expressed as target uris:
  //cell/jobs/gen-config, because we make an action class choice on the
  entrypoint
  Non-tullia dependencies, in turn, must use the action uri, eg.
  //cell/oci-images/node:copy-to-registry because we don't have an
  action class selector for them

I think the implementation could be _greatly_ simplified, if we'd
implement the DAG features in std and add the isolation via a bundler
interface of the tullia repo. This is a lot cleaner, and if I'm not
mistaken @garbas mentioned something once about "special" type of eval
caching through the bundler interface now (or maybe in the future).