facts preset: change facts directory
When using nomad's exec driver the directory for a template file
is created as the user that nomad is running as (e.g. root),
even if uid and gid options are set on the template,
and the write bit is not set in the directory permissions.
As we are by default running as nobody inside the container
this means we cannot create new files or directories inside
a folder that is a parent of a template file.
Hence we need to change /alloc/tullia to something else
so that the github template (and conceptually also others)
can still create new files under that path.
/alloc/cicero is somewhat arbitrarily chosen to still make some sense.